Added additional tests and security fixes

+++ app/Game/Kingdoms/Controllers/KingdomsController.php
@@ -5,19 +5,11 @@ namespace App\Game\Kingdoms\Controllers;
 use App\Flare\Models\Character;
 use App\Flare\Models\KingdomLog;
 use App\Game\Kingdoms\Events\UpdateKingdomLogs;
-use App\Game\Kingdoms\Service\KingdomLogService;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 class KingdomsController extends Controller
 {
-    private $kingdomLogService;
-
-    public function __construct(KingdomLogService $kingdomLogService)
-    {
-        $this->kingdomLogService = $kingdomLogService;
-    }
-
     public function attackLogs(Character $character)
     {
         $logs = $character->kingdomAttackLogs()->where('published', true)->get();
@@ -34,7 +26,7 @@ class KingdomsController extends Controller
 
     public function batchDeleteLogs(Request $request, Character $character)
     {
-        $logs = KingdomLog::findMany($request->logs);
+        $logs = $character->kingdomAttackLogs()->whereIn('id', $request->logs)->get();
 
         if ($logs->isEmpty()) {
             return redirect()->back()->with('error', 'No log exists for your selection.');
@@ -51,6 +43,10 @@ class KingdomsController extends Controller
 
     public function deleteLog(Character $character, KingdomLog $kingdomLog)
     {
+        if ($kingdomLog->character_id !== $character->id) {
+            return redirect()->back()->with('error', 'No log exists for your selection.');
+        }
+
         $name = $kingdomLog->status;
 
         $kingdomLog->delete();
